created connection to firebasedb
