+++ src/environments/environment.ts
@@ -1,0 +1,11 @@
+export const environment = {
+  firebase: {
+    projectId: 'vbscores-6e6d1',
+    appId: '1:751083793841:web:462ed2aa0731d88e49a190',
+    storageBucket: 'vbscores-6e6d1.appspot.com',
+    apiKey: 'AIzaSyAiEpMfjn15D6XRw8MpLoldMTvR5H36qzk',
+    authDomain: 'vbscores-6e6d1.firebaseapp.com',
+    messagingSenderId: '751083793841',
+    databaseURL: "https://vbscores-6e6d1.firebase.io",
+  },
+};
